fix: compare windowed inputs at the precision they are stored in
MEDS stores `numeric_value` as float32, but the windowed grid cast each
dependency back up to float64. A charted 0.1 mcg/kg/min is cast to
0.10000000149 and wrongly judged against a genuine float64 0.1. This
lead to incorrect socring of SOFA cardiovascular as 4 where it should be 3.

+++ src/open_icu/steps/concept/transformer/windowed.py
@@ -458,7 +458,7 @@ class WindowedConceptTransformer(BaseConceptTransformer):
                         schema={
                             _SUBJECT: pl.Int64,
                             _TIME: pl.Datetime(time_unit="us"),
-                            name: pl.Float64,
+                            name: pl.Float32,
                             _event(name): pl.Int32,
                         }
                     )
@@ -468,7 +468,7 @@ class WindowedConceptTransformer(BaseConceptTransformer):
                 lf.select(
                     pl.col(_SUBJECT).cast(pl.Int64),
                     pl.col(_TIME).cast(pl.Datetime(time_unit="us")),
-                    pl.col("numeric_value").cast(pl.Float64).alias(name),
+                    pl.col("numeric_value").cast(pl.Float32).alias(name),
                     pl.lit(1, dtype=pl.Int32).alias(_event(name)),
                 )
             )
